Refactor and rename to transpile_py310.py
- Clean up and refactor code for better maintainability
- Rename from strip_types.py to transpile_py310.py to better reflect purpose
- Organize code into logical sections with clear separation of concerns
- Remove debug comments and improve documentation
- Maintain all functionality: type stripping, comment removal, docstring removal
- Add comprehensive error handling and better function naming
